Move test helpers into a separate C++ file
